Rename `required_fields` to `is_required`
