+++ base/src/test/proto/spine/test/validate/required_fields_test.proto
@@ -1,0 +1,12 @@
+syntax = "proto3";
+
+package spine.test.validate;
+
+import "spine/options.proto";
+
+option (type_url_prefix) = "type.spine.io";
+option java_package = "io.spine.test.validate";
+option java_outer_classname = "RequiredFieldsTestProto";
+option java_multiple_files = true;
+
+
