+++ tools/validation-generator/src/main/java/io/spine/tools/validate/ValidationCodeGenerator.java
@@ -50,11 +50,11 @@ import io.spine.validate.Validate;
 import io.spine.validate.option.DistinctConstraint;
 import io.spine.validate.option.GoesConstraint;
 import io.spine.validate.option.IfInvalid;
+import io.spine.validate.option.IsRequiredConstraint;
 import io.spine.validate.option.PatternConstraint;
 import io.spine.validate.option.RangedConstraint;
 import io.spine.validate.option.RequiredConstraint;
 import io.spine.validate.option.RequiredFieldConstraint;
-import io.spine.validate.option.RequiredFieldsConstraint;
 import io.spine.validate.option.ValidateConstraint;
 
 import java.lang.reflect.Type;
@@ -335,7 +335,7 @@ final class ValidationCodeGenerator implements ConstraintTranslator<Set<ClassMem
     }
 
     @Override
-    public void visitRequiredOneof(RequiredFieldsConstraint constraint) {
+    public void visitRequiredOneof(IsRequiredConstraint constraint) {
         ImmutableList<FieldDeclaration> fields = constraint.fields();
         BooleanExpression fieldsAreSet = fields
                 .stream()
